Update index.js
Added ability to return null from arrowIcon to not render anything.

use case:
I want to render a custom arrow/icon conditionally.  Before this change if I would return null, it would render the default arrow icon, regardless of the `item.hasArrowNav` prop).

Now 
if `arrowIcon` is null and `item.hasArrowNav` is false, it will render nothing. 
if `arrowIcon` is null and `item.hasArrowNav` is true, it will render the default nav arrow. 
if `arrowIcon` is *not* null and `item.hasArrowNav` is true, it will render the custom arrow.
if `arrowIcon` is *not* null and `item.hasArrowNav` is false, it will render the custom arrow.