Pass ref from Header and Item onto components
Currently, we cannot use refs with SettingsList.Header and
SettingsList.Item because they are not true components with render
methods. This commit adds itemRef and headerRef props that are passed
onto the components that Item/Header get rendered into. (Frustratingly,
we have to use the names itemRef/headerRef; `ref` is an illegal React
prop name, so we cannot access a ref function as a prop, and because
SettingsList.Header and SettingsList.Item never make it to the render
stage, we cannot access the ref function through `this.refs` either.)